pyproject.toml
[project]
name = "tiniego"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.118.0,<0.119.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "supabase (>=2.21.1,<3.0.0)",
    "tencentcloud-sdk-python-ai3d (>=3.0.1470,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
